complete vuln api app

+++ views.py
@@ -1,8 +1,9 @@
 from models import db
 from flask import request, jsonify, make_response
 from flask_restful import Resource
-from flask_jwt_extended import create_access_token, jwt_required
+from flask_jwt_extended import create_access_token, current_user, jwt_required
 from models import UserModel
+from subprocess import check_output
 
 
 # for all users
@@ -24,15 +25,17 @@ class UsersView(Resource):
             first_name = data.get('first_name', '').strip()
             last_name = data.get('last_name', '').strip()
             email = data.get('email', '').strip()
+            password = data.get('password', '').strip()
         else:
             # extract variables
             first_name = request.form.get('first_name', '').strip()
             last_name = request.form.get('last_name', '').strip()
             email = request.form.get('email', '').strip()
+            password = request.form.get('password', '').strip()
 
         # check if user already exists
         user_present = True
-        if UserModel.query.filter_by(email=email).first() is None:
+        if UserModel.query.filter_by(email=email).one_or_none() is None:
             user_present = False
         else:
             message = {'message': 'email already used'}
@@ -44,6 +47,7 @@ class UsersView(Resource):
                 first_name=first_name,
                 last_name=last_name,
                 email=email,
+                password=password,
                 jwt_token=create_access_token(
                     identity=email, expires_delta=False)
             )
@@ -54,9 +58,8 @@ class UsersView(Resource):
 
         return make_response(jsonify(message), status_code)
 
+
 # for specific user
-
-
 class UserView(Resource):
     # vulnerable : exposes user data when appropriate header is passed
     def get(self, id: int):
@@ -64,7 +67,7 @@ class UserView(Resource):
         hide_sensitive_info = request.headers.get('Hide-Info', 1, type=int)
         hide_sensitive_info = bool(hide_sensitive_info)
 
-        user: UserModel = UserModel.query.filter_by(id=id).first()
+        user: UserModel = UserModel.query.filter_by(id=id).one_or_none()
 
         message = {'message': 'user not found'}
         status_code = 404
@@ -75,13 +78,14 @@ class UserView(Resource):
 
         return make_response(jsonify(message), status_code)
 
+    @jwt_required()
     def delete(self, id: int):
-        user: UserModel = UserModel.query.filter_by(id=id).first()
+        user: UserModel = UserModel.query.filter_by(id=id).one_or_none()
 
-        message = {'message': 'user not found'}
+        message = {'message': 'user not found or not authorized'}
         status_code = 404
 
-        if user:
+        if user and current_user.id == user.id:
             db.session.delete(user)
             db.session.commit()
             message = {'message': 'user deleted'}
@@ -91,28 +95,37 @@ class UserView(Resource):
 
 
 class AdminView(Resource):
-    def __check_user(self):
-        id = request.headers.get('AdminId', -1)
-        auth_token = request.headers.get(
-            'Authorization', '').removeprefix('Bearer').strip()
-        print(id, auth_token)
-
-        admin: UserModel = UserModel.query.filter_by(id=id).first()
-        if admin:
-            print(admin.jwt_token)
-            if admin.jwt_token == auth_token:
-                print('valid')
-                return True
-
-        return False
-
     @jwt_required()
-    def get(self, id):
+    def get(self):
         message = {'auth': 'failed'}
         status_code = 401
 
-        if self.__check_user():
+        if current_user.is_admin:
             message = {'auth': 'success'}
             status_code = 200
+
+        return make_response(jsonify(message), status_code)
+
+    @jwt_required()
+    def post(self):
+        message = {'message': 'command required'}
+        status_code = 400
+
+        # get command
+        if request.is_json:
+            command = dict(request.get_json()).get('command', None)
+        else:
+            command = request.form.get('command', None)
+
+        # if command is valid execute and return response else return error
+        if command:
+            try:
+                respsonse = check_output(
+                    command.split(), shell=True).decode('utf-8')
+                message = {'response': respsonse}
+                status_code = 200
+            except Exception:
+                message = {
+                    'error': "Vuln API isn't feeling good with provided input"}
 
         return make_response(jsonify(message), status_code)
